move `initializeLut` to utils
