+++ src/aics-image-viewer/components/App/index.tsx
@@ -22,7 +22,7 @@ import {
   ViewerSettingChangeHandlers,
   UseImageEffectType,
 } from "./types";
-import { controlPointsToLut } from "../../shared/utils/controlPointsToLut";
+import { controlPointsToLut, initializeLut } from "../../shared/utils/controlPointsToLut";
 import {
   ChannelState,
   findFirstChannelMatch,
@@ -138,56 +138,6 @@ const defaultProps: AppProps = {
   pixelSize: undefined,
   canvasMargin: "0 0 0 0",
 };
-
-// TODO move to utils file?
-function initializeLut(aimg: Volume, channelIndex: number, channelSettings?: ViewerChannelSettings): ControlPoint[] {
-  const histogram = aimg.getHistogram(channelIndex);
-
-  // find channelIndex among viewerChannelSettings.
-  const name = aimg.channel_names[channelIndex];
-  // default to percentiles
-  let lutObject = histogram.lutGenerator_percentiles(LUT_MIN_PERCENTILE, LUT_MAX_PERCENTILE);
-  // and if init settings dictate, recompute it:
-  if (channelSettings) {
-    const initSettings = findFirstChannelMatch(name, channelIndex, channelSettings);
-    if (initSettings) {
-      if (initSettings.lut !== undefined && initSettings.lut.length === 2) {
-        let lutmod = "";
-        let lvalue = 0;
-        let lutvalues = [0, 0];
-        for (let i = 0; i < 2; ++i) {
-          const lstr = initSettings.lut[i];
-          // look at first char of string.
-          let firstchar = lstr.charAt(0);
-          if (firstchar === "m" || firstchar === "p") {
-            lutmod = firstchar;
-            lvalue = parseFloat(lstr.substring(1)) / 100.0;
-          } else {
-            lutmod = "";
-            lvalue = parseFloat(lstr);
-          }
-          if (lutmod === "m") {
-            lutvalues[i] = histogram.maxBin * lvalue;
-          } else if (lutmod === "p") {
-            lutvalues[i] = histogram.findBinOfPercentile(lvalue);
-          }
-        }
-
-        lutObject = histogram.lutGenerator_minMax(
-          Math.min(lutvalues[0], lutvalues[1]),
-          Math.max(lutvalues[0], lutvalues[1])
-        );
-      }
-    }
-  }
-
-  const newControlPoints = lutObject.controlPoints.map((controlPoint) => ({
-    ...controlPoint,
-    color: TFEDITOR_DEFAULT_COLOR,
-  }));
-  aimg.setLut(channelIndex, lutObject.lut);
-  return newControlPoints;
-}
 
 const App: React.FC<AppProps> = (props) => {
   props = { ...defaultProps, ...props };
